teste5

+++ src/budget.f90
@@ -410,7 +410,7 @@ contains
             germinated_seeds(ri) = nint(seed_bank(ri)*0.5) !!GERMINATION
             print *, "***** Germinaram:", germinated_seeds(ri)
 
-            seed_bank(ri) = nint(seed_bank(ri) - germinated_seeds(ri)) !!UPDATE SEEDBANK
+            seed_bank(ri) = seed_bank(ri) - germinated_seeds(ri) !!UPDATE SEEDBANK
             print *, "Tamanho do banco de sementes após a germinação:", seed_bank(ri)
 
             germinated_seeds(ri) = 0.0D0
